# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Get covid data based on your geo-location in India",
        "title": "Covid 19 API Server",
        "contact": {
            "name": "Arpit Bhatnagar",
            "email": "arpitbhatnagar10@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "paths": {
        "/healthcheck": {
            "get": {
                "description": "Check server's health",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "healthcheck"
                ],
                "summary": "Server's health-check",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.HealthCheckResponse"
                        }
                    }
                }
            }
        },
        "/v1/data/geo": {
            "get": {
                "description": "Get India \u0026 State's covid data based on geo-location",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "covidApi"
                ],
                "summary": "Get Covid Data for State",
                "parameters": [
                    {
                        "type": "number",
                        "description": "Latitude",
                        "name": "lat",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "Longitude",
                        "name": "lng",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.GeoCovidDataResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/data/refresh": {
            "post": {
                "description": "Fetches and persists India's covid 19 data in DB",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "covidApi"
                ],
                "summary": "Populate Covid19 Data",
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.DataIngestResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.CovidRegionResponse": {
            "type": "object",
            "properties": {
                "activeCases": {
                    "description": "Total Active Cases",
                    "type": "string",
                    "example": "0"
                },
                "confirmedCases": {
                    "description": "Total Confirmed Cases",
                    "type": "string",
                    "example": "0"
                },
                "deaths": {
                    "description": "Total Deaths",
                    "type": "string",
                    "example": "0"
                },
                "recovered": {
                    "description": "Total Recovered",
                    "type": "string",
                    "example": "0"
                },
                "region": {
                    "description": "Data Region",
                    "type": "string"
                },
                "remoteSyncTime": {
                    "description": "Instant at which data was updated by 3rd Party Data Provider",
                    "type": "string"
                }
            }
        },
        "models.DataIngestResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "description": "Message",
                    "type": "string"
                },
                "updatedAt": {
                    "description": "Data Updated At",
                    "type": "string"
                }
            }
        },
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "description": "Error Message",
                    "type": "string"
                },
                "status": {
                    "description": "Error",
                    "type": "string"
                }
            }
        },
        "models.GeoCovidDataResponse": {
            "type": "object",
            "properties": {
                "india": {
                    "description": "India's covid data",
                    "$ref": "#/definitions/models.CovidRegionResponse"
                },
                "last_updated_at": {
                    "description": "Data last updated at",
                    "type": "string"
                },
                "state": {
                    "description": "State's covid data",
                    "$ref": "#/definitions/models.CovidRegionResponse"
                }
            }
        },
        "models.HealthCheckResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "description": "Status of server",
                    "type": "string"
                }
            }
        }
    }
}